requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0

# RAG minimal
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0

# Utilities (opsional)
tqdm>=4.66.0